Defer downloading of images until the file data is accessed.
Also try to parsed a sha1 hash from the url, and store it as an
attribute, so the file_hash function will use it. You can use this feature
to avoid unnescessary downloading of files. Just compute the hash on the
client side and send it along in the url like this::

    http://example.com/image.jpg?sha1=751499a82438277cb3cfb5db268bd41696739b3b

It will only download if not allready available locally.